+++ Cargo.toml
@@ -14,23 +14,24 @@ rust-version = "1.69.0"
 exclude = ["src/main.rs", "*.png", "*.svg", "*.gif", "*.pkl"]
 
 [dependencies]
-nalgebra = { version = "0.33.0", features = ["serde-serialize"] }
-ctrlc = "3.4.5"
-dyn-clone = "1.0.17"
-serde = { version = "1.0.214", features = ["derive", "rc"] }
-serde-pickle = "1.2.0"
+nalgebra = { version = "0.33", features = ["serde-serialize"] }
+ctrlc = "3.4"
+dyn-clone = "1.0"
+serde = { version = "1.0", features = ["derive", "rc"] }
+serde-pickle = "1.2"
 spec_math = "0.1.6"
-fastrand = "2.2.0"
+fastrand = "2.3"
 fastrand-contrib = "0.1.0"
 logsumexp = "0.1.0"
 parking_lot = { version = "0.12.3", features = ["serde"] }
 rustfft = "6.2.0"
+num = "0.4.3"
 
 [dev-dependencies]
-criterion = { version = "2.7.2", package = "codspeed-criterion-compat", features = [
+criterion = { version = "2.10", package = "codspeed-criterion-compat", features = [
   "html_reports",
 ] }
-approx = "0.5.1"
+approx = "0.5"
 
 [features]
 default = []
