+++ src/core/abort_signal.rs
@@ -2,23 +2,7 @@ use std::sync::atomic::{AtomicBool, Ordering};
 
 use parking_lot::Once;
 
-/// A trait for abort signals.
-/// This trait is used in minimizers to check if the user has requested to abort the calculation.
-pub trait AbortSignal {
-    /// Return `true` if the user has requested to abort the calculation.
-    fn is_aborted(&self) -> bool;
-    /// Abort the calculation. Make `is_aborted()` return `true`.
-    fn abort(&self);
-    /// Reset the abort signal. Make `is_aborted()` return `false`.
-    fn reset(&self);
-    /// Return a boxed version of the signal.
-    fn boxed(self) -> Box<Self>
-    where
-        Self: Sized,
-    {
-        Box::new(self)
-    }
-}
+use crate::traits::AbortSignal;
 
 static INIT: Once = Once::new();
 static CTRL_C_PRESSED: AtomicBool = AtomicBool::new(false);
