+++ src/lib.rs
@@ -191,43 +191,27 @@
     missing_docs
 )]
 
-use std::fmt::{Debug, Display};
-
 use fastrand::Rng;
 use fastrand_contrib::RngExt;
 use nalgebra::{Cholesky, DMatrix, DVector};
-use serde::{Deserialize, Serialize};
 
 /// Module containing minimization algorithms
-pub mod algorithms;
-pub use algorithms::{Minimizer, Status};
+pub mod core;
+
+/// Module containing all traits
+pub mod traits;
 
 /// Module containing MCMC sampling algorithms
 pub mod samplers;
-pub use samplers::{Ensemble, Sampler};
 
 /// Module containing swarm algorithms
 pub mod swarms;
-pub use swarms::{Swarm, SwarmMinimizer};
 
-/// Module containing various kinds of observers
-pub mod observers;
+/// Module containing various minimization algorithms
+pub mod solvers;
+
 /// Module containing standard functions for testing algorithms
 pub mod test_functions;
-
-/// Module containing a trait for aborting algorithms
-pub mod abort_signal;
-pub use abort_signal::NopAbortSignal;
-
-/// Module containing useful traits
-pub mod traits {
-    pub use crate::abort_signal::AbortSignal;
-    pub use crate::algorithms::Algorithm;
-    pub use crate::observers::{MCMCObserver, Observer};
-    pub use crate::samplers::MCMCAlgorithm;
-    pub use crate::swarms::SwarmAlgorithm;
-    pub use crate::{Function, SampleFloat};
-}
 
 /// A floating-point number type (defaults to [`f64`], see `f32` feature).
 #[cfg(not(feature = "f32"))]
@@ -310,497 +294,6 @@ impl RandChoice for Rng {
     }
 }
 
-/// Describes a point in parameter space that can be used in [`Algorithm`](`crate::traits::Algorithm`)s.
-#[derive(PartialEq, Clone, Default, Debug, Serialize, Deserialize)]
-pub struct Point {
-    x: DVector<Float>,
-    fx: Float,
-}
-impl Point {
-    /// Getter method for the point's position
-    pub const fn get_x(&self) -> &DVector<Float> {
-        &self.x
-    }
-    /// Getter method for the point's evaluation
-    pub const fn get_fx(&self) -> Float {
-        self.fx
-    }
-    /// Get the dimension of the underlying space.
-    #[allow(clippy::len_without_is_empty)]
-    pub fn dimension(&self) -> usize {
-        self.x.len()
-    }
-    /// Convert the [`Point`] into a [`Vec`]-`Float` tuple.
-    pub fn into_vec_val(self) -> (Vec<Float>, Float) {
-        let fx = self.get_fx_checked();
-        (self.x.data.into(), fx)
-    }
-    /// Evaluate the given function at the point's coordinate and set the `fx` value to the result.
-    ///
-    /// # Errors
-    ///
-    /// Returns an `Err(E)` if the evaluation fails. Users should implement this trait to return a
-    /// `std::convert::Infallible` if the function evaluation never fails.
-    pub fn evaluate<U, E>(
-        &mut self,
-        func: &dyn Function<U, E>,
-        user_data: &mut U,
-    ) -> Result<(), E> {
-        if self.fx.is_nan() {
-            self.fx = func.evaluate(self.x.as_slice(), user_data)?;
-        }
-        Ok(())
-    }
-    /// Evaluate the given function at the point's coordinate and set the `fx` value to the result.
-    /// This function assumes `x` is an internal, unbounded vector, but performs a coordinate transform
-    /// to bound `x` when evaluating the function.
-    ///
-    /// # Errors
-    ///
-    /// Returns an `Err(E)` if the evaluation fails. Users should implement this trait to return a
-    /// `std::convert::Infallible` if the function evaluation never fails.
-    pub fn evaluate_bounded<U, E>(
-        &mut self,
-        func: &dyn Function<U, E>,
-        bounds: Option<&Vec<Bound>>,
-        user_data: &mut U,
-    ) -> Result<(), E> {
-        if self.fx.is_nan() {
-            self.fx = func.evaluate_bounded(self.x.as_slice(), bounds, user_data)?;
-        }
-        Ok(())
-    }
-    /// Compare two points by their `fx` value.
-    pub fn total_cmp(&self, other: &Self) -> std::cmp::Ordering {
-        self.fx.total_cmp(&other.fx)
-    }
-    /// Move the point to a new position, resetting the evaluation of the point
-    pub fn set_position(&mut self, x: DVector<Float>) {
-        self.x = x;
-        self.fx = Float::NAN;
-    }
-    /// Get the current evaluation of the point, if it has been evaluated
-    ///
-    /// # Panics
-    ///
-    /// This method will panic if the point is unevaluated.
-    pub fn get_fx_checked(&self) -> Float {
-        assert!(!self.fx.is_nan(), "Point value requested before evaluation");
-        self.fx
-    }
-    /// Converts the point's `x` from an unbounded space to a bounded one.
-    pub fn to_bounded(&self, bounds: Option<&Vec<Bound>>) -> Self {
-        Self {
-            x: Bound::to_bounded(self.x.as_slice(), bounds),
-            fx: self.fx,
-        }
-    }
-}
-
-impl Display for Point {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        writeln!(f, "fx: {}", self.fx)?;
-        writeln!(f, "{}", self.x)
-    }
-}
-
-impl From<DVector<Float>> for Point {
-    fn from(value: DVector<Float>) -> Self {
-        Self {
-            x: value,
-            fx: Float::NAN,
-        }
-    }
-}
-impl From<Vec<Float>> for Point {
-    fn from(value: Vec<Float>) -> Self {
-        Self {
-            x: DVector::from_vec(value),
-            fx: Float::NAN,
-        }
-    }
-}
-impl<'a> From<&'a Point> for &'a Vec<Float> {
-    fn from(value: &'a Point) -> Self {
-        value.x.data.as_vec()
-    }
-}
-impl From<&[Float]> for Point {
-    fn from(value: &[Float]) -> Self {
-        Self {
-            x: DVector::from_column_slice(value),
-            fx: Float::NAN,
-        }
-    }
-}
-impl<'a> From<&'a Point> for &'a [Float] {
-    fn from(value: &'a Point) -> Self {
-        value.x.data.as_slice()
-    }
-}
-impl PartialOrd for Point {
-    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
-        self.fx.partial_cmp(&other.fx)
-    }
-}
-
-/// An enum that describes a bound/limit on a parameter in a minimization.
-///
-/// [`Bound`]s take a generic `T` which represents some scalar numeric value. They can be used by
-/// bounded algorithms directly, or by some unbounded algorithms using parameter space
-/// transformations.
-#[derive(Default, Copy, Clone, Debug, Serialize, Deserialize)]
-pub enum Bound {
-    #[default]
-    /// `(-inf, +inf)`
-    NoBound,
-    /// `(min, +inf)`
-    LowerBound(Float),
-    /// `(-inf, max)`
-    UpperBound(Float),
-    /// `(min, max)`
-    LowerAndUpperBound(Float, Float),
-}
-impl Display for Bound {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "({}, {})", self.lower(), self.upper())
-    }
-}
-impl From<(Float, Float)> for Bound {
-    fn from(value: (Float, Float)) -> Self {
-        assert!(value.0 < value.1);
-        match (value.0.is_finite(), value.1.is_finite()) {
-            (true, true) => Self::LowerAndUpperBound(value.0, value.1),
-            (true, false) => Self::LowerBound(value.0),
-            (false, true) => Self::UpperBound(value.1),
-            (false, false) => Self::NoBound,
-        }
-    }
-}
-impl From<(Option<Float>, Option<Float>)> for Bound {
-    fn from(value: (Option<Float>, Option<Float>)) -> Self {
-        assert!(value.0 < value.1);
-        match (value.0, value.1) {
-            (Some(lb), Some(ub)) => Self::LowerAndUpperBound(lb, ub),
-            (Some(lb), None) => Self::LowerBound(lb),
-            (None, Some(ub)) => Self::UpperBound(ub),
-            (None, None) => Self::NoBound,
-        }
-    }
-}
-impl Bound {
-    /// Get a value in the uniform distribution between `lower` and `upper`.
-    #[cfg(not(feature = "f32"))]
-    pub fn get_uniform(&self, rng: &mut Rng) -> Float {
-        rng.f64_range(self.lower()..self.upper()) as Float
-    }
-    /// Get a value in the uniform distribution between `lower` and `upper`.
-    #[cfg(feature = "f32")]
-    pub fn get_uniform(&self, rng: &mut Rng) -> Float {
-        rng.f32_range(self.lower()..self.upper()) as Float
-    }
-    /// Checks whether the given `value` is compatible with the bounds.
-    pub fn contains(&self, value: Float) -> bool {
-        match self {
-            Self::NoBound => true,
-            Self::LowerBound(lb) => value >= *lb,
-            Self::UpperBound(ub) => value <= *ub,
-            Self::LowerAndUpperBound(lb, ub) => value >= *lb && value <= *ub,
-        }
-    }
-    /// Checks whether the given [`DVector`] is compatible with the list of bounds in each
-    /// coordinate.
-    pub fn contains_vec(bounds: &[Self], vec: &DVector<Float>) -> bool {
-        for (bound, value) in bounds.iter().zip(vec) {
-            if !bound.contains(*value) {
-                return false;
-            }
-        }
-        true
-    }
-    /// Checks whether the given `value` is compatible with the bound and returns `0.0` if it is,
-    /// and the distance to the bound otherwise signed by whether the bound is a lower (`-`) or
-    /// upper (`+`) bound.
-    pub fn bound_excess(&self, value: Float) -> Float {
-        match self {
-            Self::NoBound => 0.0,
-            Self::LowerBound(lb) => {
-                if value >= *lb {
-                    0.0
-                } else {
-                    value - lb
-                }
-            }
-            Self::UpperBound(ub) => {
-                if value <= *ub {
-                    0.0
-                } else {
-                    value - ub
-                }
-            }
-            Self::LowerAndUpperBound(lb, ub) => {
-                if value < *lb {
-                    value - lb
-                } else if value > *ub {
-                    value - ub
-                } else {
-                    0.0
-                }
-            }
-        }
-    }
-    /// Checks whether each of the given [`DVector`]'s coordinates are compatible with the bounds
-    /// and returns a [`DVector`] containing the result of [`Bound::bound_excess`] at each
-    /// coordinate.
-    pub fn bounds_excess(bounds: &[Self], vec: &DVector<Float>) -> DVector<Float> {
-        bounds
-            .iter()
-            .zip(vec)
-            .map(|(b, v)| b.bound_excess(*v))
-            .collect::<Vec<Float>>()
-            .into()
-    }
-    /// Returns the lower bound or `-inf` if there is none.
-    pub const fn lower(&self) -> Float {
-        match self {
-            Self::NoBound => Float::NEG_INFINITY,
-            Self::LowerBound(lb) => *lb,
-            Self::UpperBound(_) => Float::NEG_INFINITY,
-            Self::LowerAndUpperBound(lb, _) => *lb,
-        }
-    }
-    /// Returns the upper bound or `+inf` if there is none.
-    pub const fn upper(&self) -> Float {
-        match self {
-            Self::NoBound => Float::INFINITY,
-            Self::LowerBound(_) => Float::INFINITY,
-            Self::UpperBound(ub) => *ub,
-            Self::LowerAndUpperBound(_, ub) => *ub,
-        }
-    }
-    /// Checks if the given value is equal to one of the bounds.
-    ///
-    /// TODO: his just does equality comparison right now, which probably needs to be improved
-    /// to something with an epsilon (significant but not critical to most fits right now).
-    pub fn at_bound(&self, value: Float) -> bool {
-        match self {
-            Self::NoBound => false,
-            Self::LowerBound(lb) => value == *lb,
-            Self::UpperBound(ub) => value == *ub,
-            Self::LowerAndUpperBound(lb, ub) => value == *lb || value == *ub,
-        }
-    }
-    /// Converts an unbounded "external" parameter into a bounded "internal" one via the transform:
-    ///
-    /// Upper and lower bounds:
-    /// ```math
-    /// x_\text{int} = \arcsin\left(2\frac{x_\text{ext} - x_\text{min}}{x_\text{max} - x_\text{min}} - 1\right)
-    /// ```
-    /// Upper bound only:
-    /// ```math
-    /// x_\text{int} = \sqrt{(x_\text{max} - x_\text{ext} + 1)^2 - 1}
-    /// ```
-    /// Lower bound only:
-    /// ```math
-    /// x_\text{int} = \sqrt{(x_\text{ext} - x_\text{min} + 1)^2 - 1}
-    /// ```
-    pub fn to_bounded(values: &[Float], bounds: Option<&Vec<Self>>) -> DVector<Float> {
-        bounds
-            .map_or_else(
-                || values.to_vec(),
-                |bounds| {
-                    values
-                        .iter()
-                        .zip(bounds)
-                        .map(|(val, bound)| bound._to_bounded(*val))
-                        .collect()
-                },
-            )
-            .into()
-    }
-    fn _to_bounded(&self, val: Float) -> Float {
-        match *self {
-            Self::LowerBound(lb) => lb - 1.0 + Float::sqrt(Float::powi(val, 2) + 1.0),
-            Self::UpperBound(ub) => ub + 1.0 - Float::sqrt(Float::powi(val, 2) + 1.0),
-            Self::LowerAndUpperBound(lb, ub) => lb + (Float::sin(val) + 1.0) * (ub - lb) / 2.0,
-            Self::NoBound => val,
-        }
-    }
-    /// Converts a bounded "internal" parameter into an unbounded "external" one via the transform:
-    ///
-    /// Upper and lower bounds:
-    /// ```math
-    /// x_\text{ext} = x_\text{min} + \left(\sin(x_\text{int}) + 1\right)\frac{x_\text{max} - x_\text{min}}{2}
-    /// ```
-    /// Upper bound only:
-    /// ```math
-    /// x_\text{ext} = x_\text{max} + 1 - \sqrt{x_\text{int}^2 + 1}
-    /// ```
-    /// Lower bound only:
-    /// ```math
-    /// x_\text{ext} = x_\text{min} - 1 + \sqrt{x_\text{int}^2 + 1}
-    /// ```
-    pub fn to_unbounded(values: &[Float], bounds: Option<&Vec<Self>>) -> DVector<Float> {
-        bounds
-            .map_or_else(
-                || values.to_vec(),
-                |bounds| {
-                    values
-                        .iter()
-                        .zip(bounds)
-                        .map(|(val, bound)| bound._to_unbounded(*val))
-                        .collect()
-                },
-            )
-            .into()
-    }
-    fn _to_unbounded(&self, val: Float) -> Float {
-        match *self {
-            Self::LowerBound(lb) => Float::sqrt(Float::powi(val - lb + 1.0, 2) - 1.0),
-            Self::UpperBound(ub) => Float::sqrt(Float::powi(ub - val + 1.0, 2) - 1.0),
-            Self::LowerAndUpperBound(lb, ub) => Float::asin(2.0 * (val - lb) / (ub - lb) - 1.0),
-            Self::NoBound => val,
-        }
-    }
-}
-
-/// A trait which describes a function $`f(\mathbb{R}^n) \to \mathbb{R}`$
-///
-/// Such a function may also take a `user_data: &mut U` field which can be used to pass external
-/// arguments to the function during minimization, or can be modified by the function itself.
-///
-/// The `Function` trait takes a generic `T` which represents a numeric scalar, a generic `U`
-/// representing the type of user data/arguments, and a generic `E` representing any possible
-/// errors that might be returned during function execution.
-///
-/// There is also a default implementation of a gradient function which uses a central
-/// finite-difference method to evaluate derivatives. If an exact gradient is known, it can be used
-/// to speed up gradient-dependent algorithms.
-pub trait Function<U, E> {
-    /// The evaluation of the function at a point `x` with the given arguments/user data.
-    ///
-    /// # Errors
-    ///
-    /// Returns an `Err(E)` if the evaluation fails. Users should implement this trait to return a
-    /// `std::convert::Infallible` if the function evaluation never fails.
-    fn evaluate(&self, x: &[Float], user_data: &mut U) -> Result<Float, E>;
-    /// The evaluation of the function at a point `x` with the given arguments/user data. This
-    /// function assumes `x` is an internal, unbounded vector, but performs a coordinate transform
-    /// to bound `x` when evaluating the function.
-    ///
-    /// # Errors
-    ///
-    /// Returns an `Err(E)` if the evaluation fails. Users should implement this trait to return a
-    /// `std::convert::Infallible` if the function evaluation never fails.
-    fn evaluate_bounded(
-        &self,
-        x: &[Float],
-        bounds: Option<&Vec<Bound>>,
-        user_data: &mut U,
-    ) -> Result<Float, E> {
-        self.evaluate(Bound::to_bounded(x, bounds).as_slice(), user_data)
-    }
-    /// The evaluation of the gradient at a point `x` with the given arguments/user data.
-    ///
-    /// # Errors
-    ///
-    /// Returns an `Err(E)` if the evaluation fails. See [`Function::evaluate`] for more
-    /// information.
-    fn gradient(&self, x: &[Float], user_data: &mut U) -> Result<DVector<Float>, E> {
-        let n = x.len();
-        let x = DVector::from_column_slice(x);
-        let mut grad = DVector::zeros(n);
-        // This is technically the best step size for the gradient, cbrt(eps) * x_i (or just
-        // cbrt(eps) if x_i = 0)
-        let h: DVector<Float> = x
-            .iter()
-            .map(|&xi| Float::cbrt(Float::EPSILON) * (xi.abs() + 1.0))
-            .collect::<Vec<_>>()
-            .into();
-        for i in 0..n {
-            let mut x_plus = x.clone();
-            let mut x_minus = x.clone();
-            x_plus[i] += h[i];
-            x_minus[i] -= h[i];
-            let f_plus = self.evaluate(x_plus.as_slice(), user_data)?;
-            let f_minus = self.evaluate(x_minus.as_slice(), user_data)?;
-            grad[i] = (f_plus - f_minus) / (2.0 * h[i]);
-        }
-        Ok(grad)
-    }
-    /// The evaluation of the gradient at a point `x` with the given arguments/user data. This
-    /// function assumes `x` is an internal, unbounded vector, but performs a coordinate transform
-    /// to bound `x` when evaluating the function.
-    ///
-    /// # Errors
-    ///
-    /// Returns an `Err(E)` if the evaluation fails. See [`Function::evaluate`] for more
-    /// information.
-    fn gradient_bounded(
-        &self,
-        x: &[Float],
-        bounds: Option<&Vec<Bound>>,
-        user_data: &mut U,
-    ) -> Result<DVector<Float>, E> {
-        self.gradient(Bound::to_bounded(x, bounds).as_slice(), user_data)
-    }
-
-    /// The evaluation of the hessian at a point `x` with the given arguments/user data.
-    ///
-    /// # Errors
-    ///
-    /// Returns an `Err(E)` if the evaluation fails. See [`Function::evaluate`] for more
-    /// information.
-    fn hessian(&self, x: &[Float], user_data: &mut U) -> Result<DMatrix<Float>, E> {
-        let x = DVector::from_column_slice(x);
-        let h: DVector<Float> = x
-            .iter()
-            .map(|&xi| Float::cbrt(Float::EPSILON) * (xi.abs() + 1.0))
-            .collect::<Vec<_>>()
-            .into();
-        let mut res = DMatrix::zeros(x.len(), x.len());
-        let mut g_plus = DMatrix::zeros(x.len(), x.len());
-        let mut g_minus = DMatrix::zeros(x.len(), x.len());
-        // g+ and g- are such that
-        // g+[(i, j)] = g[i](x + h_je_j) and
-        // g-[(i, j)] = g[i](x - h_je_j)
-        for i in 0..x.len() {
-            let mut x_plus = x.clone();
-            let mut x_minus = x.clone();
-            x_plus[i] += h[i];
-            x_minus[i] -= h[i];
-            g_plus.set_column(i, &self.gradient(x_plus.as_slice(), user_data)?);
-            g_minus.set_column(i, &self.gradient(x_minus.as_slice(), user_data)?);
-            for j in 0..=i {
-                if i == j {
-                    res[(i, j)] = (g_plus[(i, j)] - g_minus[(i, j)]) / (2.0 * h[i]);
-                } else {
-                    res[(i, j)] = ((g_plus[(i, j)] - g_minus[(i, j)]) / (4.0 * h[j]))
-                        + ((g_plus[(j, i)] - g_minus[(j, i)]) / (4.0 * h[i]));
-                    res[(j, i)] = res[(i, j)];
-                }
-            }
-        }
-        Ok(res)
-    }
-    /// The evaluation of the hessian at a point `x` with the given arguments/user data. This
-    /// function assumes `x` is an internal, unbounded vector, but performs a coordinate transform
-    /// to bound `x` when evaluating the function.
-    ///
-    /// # Errors
-    ///
-    /// Returns an `Err(E)` if the evaluation fails. See [`Function::evaluate`] for more
-    /// information.
-    fn hessian_bounded(
-        &self,
-        x: &[Float],
-        bounds: Option<&Vec<Bound>>,
-        user_data: &mut U,
-    ) -> Result<DMatrix<Float>, E> {
-        self.hessian(Bound::to_bounded(x, bounds).as_slice(), user_data)
-    }
-}
-
 pub(crate) fn generate_random_vector(
     dimension: usize,
     lb: Float,
@@ -824,13 +317,13 @@ pub(crate) fn generate_random_vector_in_limits(
 mod tests {
     use std::convert::Infallible;
 
-    use crate::algorithms::{Minimizer, LBFGSB};
+    use crate::{core::Problem, solvers::lbfgsb::LBFGSB};
 
     #[test]
     #[allow(unused_variables)]
-    fn test_minimizer_constructor() {
+    fn test_problem_constructor() {
         #[allow(clippy::box_default)]
         let algo: LBFGSB<(), Infallible> = LBFGSB::default();
-        let minimizer = Minimizer::new(Box::new(algo), 5);
+        let problem = Problem::new(Box::new(algo), 2);
     }
 }
