+++ src/samplers/mod.rs
@@ -18,8 +18,9 @@ use rustfft::FftPlanner;
 use serde::{Deserialize, Serialize};
 
 use crate::{
-    traits::{AbortSignal, MCMCObserver},
-    Float, Function, Point,
+    core::Point,
+    traits::{AbortSignal, CostFunction, MCMCObserver},
+    Float,
 };
 
 /// A MCMC walker containing a history of past samples
@@ -65,7 +66,7 @@ impl Walker {
     /// information.
     pub fn evaluate_latest<U, E>(
         &mut self,
-        func: &dyn Function<U, E>,
+        func: &dyn CostFunction<U, E>,
         user_data: &mut U,
     ) -> Result<(), E> {
         self.get_latest().write().evaluate(func, user_data)
@@ -133,7 +134,7 @@ impl Ensemble {
     /// information.
     pub fn evaluate_latest<U, E>(
         &mut self,
-        func: &dyn Function<U, E>,
+        func: &dyn CostFunction<U, E>,
         user_data: &mut U,
     ) -> Result<(), E> {
         for walker in self.walkers.iter_mut() {
@@ -356,11 +357,11 @@ pub trait MCMCAlgorithm<U, E> {
     ///
     /// # Errors
     ///
-    /// Returns an `Err(E)` if the evaluation fails. See [`Function::evaluate`] for more
+    /// Returns an `Err(E)` if the evaluation fails. See [`CostFunction::evaluate`] for more
     /// information.
     fn initialize(
         &mut self,
-        func: &dyn Function<U, E>,
+        func: &dyn CostFunction<U, E>,
         user_data: &mut U,
         ensemble: &mut Ensemble,
     ) -> Result<(), E>;
@@ -369,12 +370,12 @@ pub trait MCMCAlgorithm<U, E> {
     ///
     /// # Errors
     ///
-    /// Returns an `Err(E)` if the evaluation fails. See [`Function::evaluate`] for more
+    /// Returns an `Err(E)` if the evaluation fails. See [`CostFunction::evaluate`] for more
     /// information.
     fn step(
         &mut self,
         i_step: usize,
-        func: &dyn Function<U, E>,
+        func: &dyn CostFunction<U, E>,
         user_data: &mut U,
         ensemble: &mut Ensemble,
     ) -> Result<(), E>;
@@ -382,11 +383,11 @@ pub trait MCMCAlgorithm<U, E> {
     ///
     /// # Errors
     ///
-    /// Returns an `Err(E)` if the evaluation fails. See [`Function::evaluate`] for more
+    /// Returns an `Err(E)` if the evaluation fails. See [`CostFunction::evaluate`] for more
     /// information.
     fn check_for_termination(
         &mut self,
-        func: &dyn Function<U, E>,
+        func: &dyn CostFunction<U, E>,
         user_data: &mut U,
         ensemble: &mut Ensemble,
     ) -> Result<bool, E>;
@@ -394,12 +395,12 @@ pub trait MCMCAlgorithm<U, E> {
     ///
     /// # Errors
     ///
-    /// Returns an `Err(E)` if the evaluation fails. See [`Function::evaluate`] for more
+    /// Returns an `Err(E)` if the evaluation fails. See [`CostFunction::evaluate`] for more
     /// information.
     #[allow(unused_variables)]
     fn postprocessing(
         &mut self,
-        func: &dyn Function<U, E>,
+        func: &dyn CostFunction<U, E>,
         user_data: &mut U,
         ensemble: &mut Ensemble,
     ) -> Result<(), E> {
@@ -434,7 +435,7 @@ impl<U, E> Sampler<U, E> {
         self.observers.push(observer);
         self
     }
-    /// Minimize the given [`Function`] starting at the point `x0`.
+    /// Minimize the given [`CostFunction`] starting at the point `x0`.
     ///
     /// This method first runs [`MCMCAlgorithm::initialize`], then runs [`MCMCAlgorithm::step`] in a loop,
     /// terminating if [`MCMCAlgorithm::check_for_termination`] returns `true` or if
@@ -444,11 +445,11 @@ impl<U, E> Sampler<U, E> {
     ///
     /// # Errors
     ///
-    /// Returns an `Err(E)` if the evaluation fails. See [`Function::evaluate`] for more
+    /// Returns an `Err(E)` if the evaluation fails. See [`CostFunction::evaluate`] for more
     /// information.
     pub fn sample(
         &mut self,
-        func: &dyn Function<U, E>,
+        func: &dyn CostFunction<U, E>,
         user_data: &mut U,
         n_steps: usize,
         abort_signal: Box<dyn AbortSignal>,
